refactor: 약관 및 정책 페이지 내 client 컴포넌트 clientBoundary 디렉토리로 이동 (#21)
